requirements.txt
fastapi
uvicorn
pydantic
email-validator
motor
pymongo
PyJWT
passlib[bcrypt]
pytest
pytest-anyio
httpx
